Inverse of A
The inverse of a matrix A is denoted by A−1 such that the following relationship holds −

A*A^−1 = A^−1*A = 1
The inverse of a matrix does not always exist. If the determinant of the matrix is zero, then the inverse does not exist and the matrix is singular.

Inverse of a matrix in MATLAB is calculated using the inv function. The inverse of a matrix A is given by inv(A)